Samples Added and ReadMe updated accordingly. Also added property to item class to get the proper ItemId back for use with the GetItem funon

+++ Explorers/D3Explorer.cs
@@ -106,7 +106,7 @@ namespace D3DotNetAPI
 
         #region Hero
 
-        public Hero GetHero (string battleTagName, int battleTagCode, int heroId)
+        public Hero GetHero (string battleTagName, int battleTagCode, long heroId)
         {
             Hero hero;
